chore: rename source dir to source
