+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "clean": "rm -rf build",
     "deploy-storybook": "storybook-to-ghpages",
-    "build": "yarn run clean && babel --out-dir build lib",
+    "build": "yarn run clean && babel --out-dir build source",
     "dev": "start-storybook -p 6006",
     "postversion": "yarn run build && git push && git push --tags && npm publish && release",
     "build-storybook": "build-storybook"
