+++ source/platform.js
