+++ source/tip.js
